:sparkles: feat: Add calculo de saldo - em desenvolvimento

+++ src/main/java/com/nathannolacio/meusaldo/dto/TransactionResponseDTO.java
@@ -4,9 +4,10 @@ import com.nathannolacio.meusaldo.model.Account;
 import com.nathannolacio.meusaldo.model.Transaction;
 import com.nathannolacio.meusaldo.model.TransactionType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record TransactionResponseDTO(Long id, LocalDate date, String description, Double amount, String type, String accountName
+public record TransactionResponseDTO(Long id, LocalDate date, String description, BigDecimal amount, String type, String accountName
 ) {
     public TransactionResponseDTO(Transaction transaction) {
         this(
